+++ packages/components/CHANGELOG.md
@@ -20,6 +20,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
   `aria` prop to provide additional context for screen readers.
   ([#3969](https://github.com/porsche-design-system/porsche-design-system/pull/3969))
 
+### Fixed
+
+- `Select Wrapper`: fix axe-core violation when both `filter` and `description` prop are used
+  ([#4011](https://github.com/porsche-design-system/porsche-design-system/pull/4011))
+
 ## [3.29.0] - 2025-08-20
 
 ## [3.29.0-rc.5] - 2025-08-19